ignore warnings issued by Data::Dumper
Data::Dumper is used to generate textual representations of
structures, not for serializing them, so any warnings it issues
(e.g. unrecognized types, etc) do not provide actionable information
to the user, and serve only to confuse and distract.

In particular, the new perl object type isn't (yet?) handled by
Data::Dumper, which results in many warnings being output when
structures which contain them are passed to it.